Add support for preview and preprod (#117)

+++ indexer/tasks/src/multiera/multiera_metadata.rs
@@ -83,7 +83,7 @@ async fn handle_metadata(
             .map(
                 |((label, metadata), tx_id)| TransactionMetadataActiveModel {
                     tx_id: Set(*tx_id),
-                    label: Set(u64::from(label).to_le_bytes().to_vec()),
+                    label: Set(label.to_le_bytes().to_vec()),
                     payload: Set(metadata.encode_fragment().unwrap()),
                     ..Default::default()
                 },
